refactor(domain): DayOfWeek를 common.entity로 이동
- business 외 다른 도메인에서도 사용되어 독립적인 공통 enum 이라 판단 하, common package 로 이동
